+++ lib/src/data/mappers/code_mapper.dart
@@ -1,0 +1,13 @@
+import '../../domain/entities/blocks/code_block.dart';
+import 'block_mapper.dart';
+
+class CodeMapper implements BlockMapper<CodeBlock> {
+  const CodeMapper();
+
+  @override
+  String get supportedType => 'code';
+
+  @override
+  CodeBlock fromJson(Map<String, dynamic> data) =>
+      CodeBlock(code: (data['code'] as String?) ?? '');
+}